formalized/2604-22085-memanto-gaia/pyproject.toml
[project]
name = "2604-24198-memanto-gaia"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "riso", email = "ruoxue@gmail.com" }
]
requires-python = ">=3.13"
dependencies = ["gaia-lang"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/2604_24198_memanto"]

[tool.gaia]
type = "knowledge-package"
uuid = "b92c79d6-eafe-4625-a548-64442fc2a893"
